Dynamic introspection size
Change the introspection string generation in `send_introspection` from a static 512 characters static value to a dynamic one calculated on the items located in `interfaces_list`. This prevents a sudden block on introspection publication over a value that could be perceived as too arbitrary.
Add a warning, though, if the size is greater than 4KiB.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>