Add support to server-owned property unset
Three possible ways were found to implement the previously unsupported
feature of unsetting a property belonging to a server-owned interface.
The first was to send an `astarte_device_data_event_t` with a `NULL`
`bson_data` to the `data_event_callback`.
Similarly, a possibility was to send to it an
 `astarte_device_data_event_t` with the `bson_type` set as a newly
 introduced `BSON_TYPE_NULL`.
Lastly, it was considered the introduction of a new optional
`astarte_device_data_event_callback_t` called
`unset_data_event_callback` in charge of handling only unset property
operations.
The first two cases were deemed unsafe as the new behavior could cause
troubles to users that already employ the SDK and have defined a
`device_data_event_callback_t` without checking the two fields.
The latter was found a safer choice as it was not going to alter
previous behaviors of the SDK.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>